7-Zip Version 4.31

+++ 7zip/Archive/7z_C/7zItem.h
@@ -75,7 +75,7 @@ typedef struct _CArchiveDatabase
 {
   UInt32 NumPackStreams;
   CFileSize *PackSizes;
-  int *PackCRCsDefined;
+  Byte *PackCRCsDefined;
   UInt32 *PackCRCs;
   UInt32 NumFolders;
   CFolder *Folders;
